+++ src/pip/_internal/index.py
@@ -82,8 +82,8 @@ def _match_vcs_scheme(url):
 
     Returns the matched VCS scheme, or None if there's no match.
     """
-    from pip._internal.vcs import VcsSupport
-    for scheme in VcsSupport.schemes:
+    from pip._internal.vcs import vcs
+    for scheme in vcs.schemes:
         if url.lower().startswith(scheme) and url[len(scheme)] in '+:':
             return scheme
     return None
