+++ src/pip/_internal/vcs/bazaar.py
@@ -5,9 +5,8 @@ import os
 
 from pip._vendor.six.moves.urllib import parse as urllib_parse
 
-from pip._internal.download import path_to_url
-from pip._internal.utils.misc import display_path, rmtree
-from pip._internal.vcs import VersionControl, vcs
+from pip._internal.utils.misc import display_path, path_to_url, rmtree
+from pip._internal.vcs.versioncontrol import VersionControl, vcs
 
 logger = logging.getLogger(__name__)
