+++ src/pip/_internal/vcs/mercurial.py
@@ -5,10 +5,9 @@ import os
 
 from pip._vendor.six.moves import configparser
 
-from pip._internal.download import path_to_url
-from pip._internal.utils.misc import display_path
+from pip._internal.utils.misc import display_path, path_to_url
 from pip._internal.utils.temp_dir import TempDirectory
-from pip._internal.vcs import VersionControl, vcs
+from pip._internal.vcs.versioncontrol import VersionControl, vcs
 
 logger = logging.getLogger(__name__)
